+++ examples/graph-autopipe/stages/stages.go
@@ -6,10 +6,10 @@ import (
 	"io/ioutil"
 	"net/http"
 
-	"github.com/mariomac/pipes/pkg/graph/stage"
-
-	"github.com/mariomac/pipes/pkg/node"
 	"github.com/sirupsen/logrus"
+
+	"github.com/mariomac/pipes/pkg/graph/stage"
+	"github.com/mariomac/pipes/pkg/node"
 )
 
 const defaultPort = 8080