+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 # next
+* High-Level: You don't need the `nodeId` annotation. If undefined, it takes the struct name
+  by default.
 * **BREAKING CHANGE**: we don't inspect anymore array/slice types, element by element. Instead,
   use StartMultiProvider and, when we do it, TerminalMultiprovider (we will decide what to do with
   MiddleMultiProvider).